+++ PSFramework/internal/tepp/scripts/localization.ps1
@@ -1,0 +1,3 @@
+﻿Register-PSFTeppScriptblock -Name 'PSFramework-LanguageNames' -ScriptBlock {
+	[System.Globalization.CultureInfo]::GetCultures([System.Globalization.CultureTypes]::AllCultures).Name
+}