+++ PSFramework/internal/tepp/tepp-assignment.ps1
@@ -26,6 +26,10 @@ Register-PSFTeppArgumentCompleter -Command Unregister-PSFConfig -Parameter Modul
 Register-PSFTeppArgumentCompleter -Command Get-PSFLicense -Parameter Filter -Name 'PSFramework-license-name'
 #endregion License
 
+#region Localization
+Register-PSFTeppArgumentCompleter -Command Import-PSFLocalizedString -Parameter Language -Name 'PSFramework-LanguageNames'
+#endregion Localization
+
 #region Logging
 Register-PSFTeppArgumentCompleter -Command Get-PSFLoggingProvider -Parameter Name -Name 'PSFramework-logging-provider'
 Register-PSFTeppArgumentCompleter -Command Install-PSFLoggingProvider -Parameter Name -Name 'PSFramework-logging-provider'