+++ PSFramework/changelog.md
@@ -4,6 +4,8 @@
  - New: Command Get-PSFPipeline grants access to the current pipeline and all its works.
  - New: Command Get-PSFScriptblock retrieves scriptblocks from a static dictionary
  - New: Command Set-PSFScriptblock stores scriptblocks in a static dictionary
+ - New: Command Get-PSFLocalizedString retrieves localied versions of strings
+ - New: Command Import-PSFLocalizedString imports localized strings into the strings store
  - New: Logging Provider for gelf / graylog
  - Upd: PsfValidateScript can now consume stored scriptblocks
  - Upd: PsfValidateScript will now understand both $_ and $args[0]